% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMR.CTD.R
\name{beta.Hellinger}
\alias{beta.Hellinger}
\title{beta.Hellinger}
\usage{
beta.Hellinger(a1, b1, a2, b2)
}
\arguments{
\item{a1}{First shape parameter of the first beta distribution.}

\item{b1}{Second shape parameter of the first beta distribution.}

\item{a2}{First shape parameter of the second beta distribution.}

\item{b2}{Second shape parameter of the second beta distribution.}
}
\value{
A numeric value representing the squared Hellinger distance.
}
\description{
Computes the squared Hellinger distance between two beta distributions with 
parameters \code{(a1, b1)} and \code{(a2, b2)}. It is a sub function used in the beta.barycenter function.
}
