% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMR.CTD.R
\name{BMR.CTD}
\alias{BMR.CTD}
\title{BMR.CTD}
\usage{
BMR.CTD(
  orig_weights,
  orig_alphas,
  orig_betas,
  M,
  max_iter = 100,
  tol = 1e-06,
  n_sample = 10000
)
}
\arguments{
\item{orig_weights}{A numeric vector of original mixture weights.}

\item{orig_alphas}{A numeric vector of \eqn{\alpha} parameters for the original Beta components.}

\item{orig_betas}{A numeric vector of \eqn{\beta} parameters for the original Beta components.}

\item{M}{The number of components in the reduced mixture.}

\item{max_iter}{Maximum number of iterations for the algorithm. Default is 100.}

\item{tol}{Tolerance for convergence based on change in total cost. Default is 1e-6.}

\item{n_sample}{Number of samples to use for initialization. Default is 10000.}
}
\value{
A list with elements:
\item{reduced_alphas}{Estimated \eqn{\alpha} parameters of the reduced mixture.}
\item{reduced_betas}{Estimated \eqn{\beta} parameters of the reduced mixture.}
\item{reduced_weights}{Weights of the reduced mixture components.}
\item{cost_history}{Vector of objective costs at each iteration.}
\item{assignments}{Cluster assignment for each original component.}
\item{n_iter}{Number of iterations run.}
\item{converged}{Logical indicating whether convergence was achieved.}
}
\description{
performs beta mixture reduction by iteratively assigning components based on Hellinger distance 
and recomputing cluster barycenters.
}
