% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix.binom.R
\name{dmix.binom}
\alias{dmix.binom}
\title{dmix.binom}
\usage{
dmix.binom(x, size, alpha, theta)
}
\arguments{
\item{x}{A vector at which the probability mass function is evaluated.}

\item{size}{The number of trials.}

\item{alpha}{A vector of the mixing probabilities.}

\item{theta}{A vector of the probabilities of success of each component.}
}
\description{
This function computes the probability mass function points in vector x.
}
\examples{
> Compute a density function from a 3 component binomial mixture
alpha = c(.5, .1, .4)
theta = c(.1, .35, .8)
size = 25
x = rmix.binom(1000, size, alpha, theta)
dmix.binom(x,size,alpha,theta)
}
