% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix.exp.R
\name{dmix.exp}
\alias{dmix.exp}
\title{dmix.exp}
\usage{
dmix.exp(x, alpha, theta)
}
\arguments{
\item{x}{The value at which we computed the cdf.}

\item{alpha}{A vector of the mixing probabilities.}

\item{theta}{A vector of the subpopulation means.}
}
\description{
This function computes the pdf of the finite exponential mixture distribution.
}
