% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix.norm.R
\name{dmix.norm}
\alias{dmix.norm}
\title{dmix.norm}
\usage{
dmix.norm(x, alpha, mu, sigma, tol = 1e-08)
}
\arguments{
\item{x}{uni-variate mixture density at x.}

\item{alpha}{A vector of the mixing proportions.}

\item{mu}{A vector of the component means.}

\item{sigma}{A vector of the component standard deviations.}

\item{tol}{A fixed value at 1e-8}
}
\value{

}
\description{
dmix.norm
}
