% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix.pois.R
\name{dmix.pois}
\alias{dmix.pois}
\title{dmix.pois}
\usage{
dmix.pois(xx, alpha, theta)
}
\arguments{
\item{xx}{A vector of x at which the probability is computed.}

\item{alpha}{A vector of the mixing proportions.}

\item{theta}{A vector of the subpopulation means.}
}
\description{
This function computes the probability mass function of the Poisson mixture
}
