% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix.beta.R
\name{dmix_beta}
\alias{dmix_beta}
\title{dmix.beta}
\usage{
dmix_beta(x, mix_porp, alpha, beta)
}
\arguments{
\item{x}{The uni-variate mixture density at x.}

\item{mix_porp}{A vector of the mixing proportions.}

\item{alpha}{A vector of the component alphas.}

\item{beta}{A vector of the component betas.}
}
\description{
A function that computes the density of the univariate beta mixture.
}
\note{
In order to plot the mixture density together with its subpopulation density, one can use the plotmix.beta function.
}
\examples{
dmix.beta(1:10,c(.5,.5),c(2,5),c(10,10))

}
