% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emiter.norm.R
\name{emiter.norm}
\alias{emiter.norm}
\title{emiter.norm}
\usage{
emiter.norm(x, para, beta, pens, para0, k)
}
\arguments{
\item{x}{}

\item{para}{}

\item{beta}{}

\item{pens}{}

\item{para0}{}

\item{k}{}
}
\value{

}
\description{
emiter.norm
}
