% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emtest.exp.R
\name{emtest.exp}
\alias{emtest.exp}
\title{emtest.exp}
\usage{
emtest.exp(
  x,
  m0 = 1,
  C = NULL,
  inival = NULL,
  len = 10,
  niter = 50,
  tol = 1e-06,
  k = 3,
  rformat = FALSE
)
}
\arguments{
\item{x}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{m0}{The order under null hypothesis.}

\item{C}{A optional tuning parameter for the EM-test procedure.}

\item{inival}{The initial values chosen for the EM-algorithm in order to compute the PMLE under the null model.}

\item{len}{The number of initial values chosen for the EM-algorithm.}

\item{niter}{The least amount of iterations for all initial values in the EM-algorithm.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}

\item{k}{The number of EM iterations needed to obtain the EM-test statistic.}

\item{rformat}{The format of the output, rformat=T means the format is determined by R.
rformat=F means the format is determined by our default setting. When the output is
larger than 0.001, it is determined by round(output,3); When the output is less than 0.001,
it is determined by signif(output,3).}
}
\description{
This function computes the EM-test statistic and the p-value for the null hypothesis H0:m=m0.
}
