% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMtest.maxmm.pois.R
\name{EMtest.maxmm.pois}
\alias{EMtest.maxmm.pois}
\title{EMtest.maxmm.pois}
\usage{
EMtest.maxmm.pois(
  count,
  freq,
  beta.i,
  alpha0,
  theta0,
  m0,
  n.init,
  n.iter,
  tol,
  max.iter
)
}
\arguments{
\item{beta.i}{The fixed mixing proportions.}

\item{m0}{The order of the finite normal mixture model.}

\item{n.init}{A computer generated n.init initials value.}

\item{n.iter}{The number of EM iterations for each initial values. The one gained the most in likelihood will be iterative further.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}

\item{xx}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{an}{}

\item{para0}{The subpopulations parameter.}
}
\description{
This function computes the PMLE of parameters under the alternative model 
             for EM-test given a beta.i value. It is used in the EMstat.pois function.
}
