% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMtest.norm.iter.R
\name{EMtest.norm.iter}
\alias{EMtest.norm.iter}
\title{EMtest.norm.iter}
\usage{
EMtest.norm.iter(xx, nn, m0, para, sigma0, beta.i, pen.size, k)
}
\arguments{
\item{xx}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{nn}{The lenght of the input data.}

\item{m0}{The order of the finite normal mixture model.}

\item{para}{The subpopulations estimated parameter.}

\item{sigma0}{The estimated variance under the null model.}

\item{beta.i}{The fixed mixing proportions.}

\item{pen.size}{A 2-dimensions vector being the size of the penalized functions for 
the mixing proportion and the variance.}

\item{k}{The amount of EM iterations in order to obtain the EM-test statistic.}
}
\description{
This function computes the EM-algorithm in the emtest.norm function.
}
