% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMtest.norm.omega.R
\name{EMtest.norm.omega}
\alias{EMtest.norm.omega}
\title{EMtest.norm.omega}
\usage{
EMtest.norm.omega(alpi, mui, sigi, alpj, muj, sigj, tol = 1e-05)
}
\arguments{
\item{alpi}{}

\item{mui}{}

\item{sigi}{}

\item{alpj}{}

\item{muj}{}

\item{sigj}{}

\item{tol}{}
}
\description{
Computes the overlapping probability of two sub-populations of the Gaussian mixture.
}
\note{
Chen, Li and Fu (2012 Jasa). Section 3.1, replaced over1() in the previous package.
}
