% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emtest.norm.R
\name{emtest.norm}
\alias{emtest.norm}
\title{emtest.norm}
\usage{
emtest.norm(
  x,
  m0 = 1,
  pen.size = NULL,
  init.val = NULL,
  n.init = 10,
  n.iter = 50,
  tol = 1e-06,
  k = 3,
  rformat = FALSE
)
}
\arguments{
\item{x}{The input data that can either be a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{m0}{The order of the finite normal mixture model under the null hypothesis.}

\item{pen.size}{A 2-dimensions vector being the size of the penalized functions for 
the mixing proportion and the variance.}

\item{init.val}{The initial value chosen for the EM-algorithm to compute the PMLE under the null model.}

\item{n.init}{The number of initial values chosen for the EM-algorithm.}

\item{n.iter}{The least amount of iterations for all initial values in the EM-algorithm.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}

\item{k}{The amount of EM iterations in order to obtain the EM-test statistic.}

\item{rformat}{F means the format of the output is determined by our default setting. When the output is
larger than 0.001, it is determined by round(output,3); When the output is less than 0.001,
it is determined by signif(output,3).}
}
\value{
Return an object of class EM-test with the following elements:
The MLE of the parameters under the null hypothesis (order = m0)
The PMLE of the parameters under the specific alternative hypothesis whose order is 2m0
EM-test statistic
P-value
Level of penalty
The number of iterations
}
\description{
emtest.norm
}
\examples{
data(grains)
out <- emtest.norm(unlist(grains),m0 = 2)
plotmix.norm(unlist(grains), alpha = out[[1]][1,], mu = out[[1]][2,], sigma = out[[1]][3,], m0 = 2)
}
\references{
Chen, J. and Li, P. (2009). Hypothesis test for normal mixture models: The EM approach. The Annals of Statistics. 37, 2523-2542.
Chen, J., Li, P. and Fu, Y. (2012). Inference on the order of a normal mixture. JASA. 107, 1096-1105.
}
\seealso{
plotmix.norm, plotmix.norm.bin, pmle.norm, rmix.norm
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
