% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emtest.norm0.R
\name{emtest.norm0}
\alias{emtest.norm0}
\title{emtest.norm0}
\usage{
emtest.norm0(
  x,
  var.sub = 1,
  m0 = 1,
  init.val = NULL,
  n.init = 10,
  n.iter = 50,
  max.iter = 5000,
  tol = 1e-06,
  k = 3,
  rformat = FALSE
)
}
\arguments{
\item{x}{The input data that can either be a vector or a matrix with the 1st column being the observed values and the 2nd column being the corresponding frequency.}

\item{var.sub}{The known component variance.}

\item{m0}{The order under the null hypothesis.}

\item{init.val}{The initial values chosen for the EM-algorithm to compute the PMLE under the null model}

\item{n.init}{The number of initial values chosen for the EM-algorithm.}

\item{n.iter}{The least amount of iterations for all initial values in the EM-algorithm.}

\item{max.iter}{The maximum number of iterations.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}

\item{k}{The amount of EM iterations in order to obtain the EM-test statistic.}

\item{rformat}{F means the format of the output is determined by our default setting. When the output is
larger than 0.001, it is determined by round(output,3); When the output is less than 0.001,
it is determined by signif(output,3).}
}
\value{
Return an object of class EM-test with the following elements:
The MLE of the parameters under the null hypothesis (order = m0)
The PMLE of the parameters under the specific alternative hypothesis whose order is 2m0
EM-test statistic
P-value
Level of penalty
The number of iterations
}
\description{
This function computes the EM-test statistic and the p-value for the hypothesis H0:m=m0.
}
\examples{
x <- rmix.norm(200,c(0.3,0.7),c(-1,2),c(1,1))
emtest.norm0(x)
}
\references{
Chen, J. and Li, P. (2009). Hypothesis test for normal mixture models: The EM approach. The Annals of Statistics. 37, 2523-2542.
Chen, J., Li, P. and Fu, Y. (2012). Inference on the order of a normal mixture. JASA. 107, 1096-1105.
}
\seealso{
plotmix.norm0, pmle.norm0, rmix.norm
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
