% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emtest.pois.R
\name{emtest.pois}
\alias{emtest.pois}
\title{emtest.pois}
\usage{
emtest.pois(
  x,
  m0 = 1,
  CC = NULL,
  init.val = NULL,
  n.init = 10,
  n.iter = 50,
  tol = 1e-06,
  k = 3,
  rformat = F,
  max.iter = 5000
)
}
\arguments{
\item{x}{The input data that can either be a vector or a matrix of observations with its 1st column being counts and 2nd column being frequencies.}

\item{m0}{The order of the finite mixture model.}

\item{CC}{The optional tuning parameter for the EM-test procedure.}

\item{init.val}{A user provided initial values for the EM-algorithm to 
compute the PMLE under the null model.}

\item{n.init}{The number of initial values chosen for the EM-algorithm.}

\item{n.iter}{The least amount of iterations for all initial values in the EM-algorithm.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}

\item{k}{The number of EM iterations required in order to obtain the EM-test statistic.}

\item{rformat}{A specific format, please see rousignif.R function.}

\item{max.iter}{The maximum amount of iterations.}
}
\value{
The function returns an object of class EM-test with the following elements:
The MLE of the parameters under the null hypothesis (order = m0)
The PMLE of the parameters under the specific alternative hypothesis whose order is 2m0
EM-test statistic
P-value
Level of penalty
The number of iterations
}
\description{
This function computes the EM-test statistic and the p-value under the null hypothesis with order m = m0.
}
\references{
Chen, J. and Li, P. (2011). Tuning the EM-test for the order of finite mixture models. The Canadian Journal of Statistics. 39, 389-404.
Li, P. and Chen, J. (2010). Testing the order of a finite mixture model. JASA. 105, 1084-1092.
Li, P., Chen, J. and Marriott, P. (2009). Non-finite Fisher information and homogeneity: The EM approach. Biometrika. 96, 411-426.
}
\seealso{
plotmix.pois, pmle.pois, rmix.pois
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
