% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emtest.thm3.R
\name{emtest.thm3}
\alias{emtest.thm3}
\title{emtest.thm3}
\usage{
emtest.thm3(tb, N = 10000, tol = 1e-08)
}
\arguments{
\item{tb}{The input data.}

\item{N}{Number of simulation.}

\item{tol}{The tolerance value.}
}
\description{
This function computes a_h as in Theorem 3 of Li and Chen (JASA2009) by MC.
It is used in the emtest.exp function.
}
