% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMR.CTD.R
\name{init_reduced_beta_mixture}
\alias{init_reduced_beta_mixture}
\title{init_reduced_beta_mixture}
\usage{
init_reduced_beta_mixture(weights, alphas, betas, M, n_sample = 10000)
}
\arguments{
\item{weights}{A numeric vector of mixture weights.}

\item{alphas}{A numeric vector of \eqn{\alpha} parameters for each Beta component.}

\item{betas}{A numeric vector of \eqn{\beta} parameters for each Beta component.}

\item{M}{The number of components in the reduced mixture.}

\item{n_sample}{The number of samples to draw from the original mixture. Default is 10000.}
}
\value{
A list containing the initialized \code{alphas}, \code{betas}, and \code{weights} 
        for the reduced Beta mixture.
}
\description{
Generates initial parameters for a reduced beta mixture by sampling from the original 
mixture and clustering the samples with k-means. The Score-Adjusted Moment estimation of the clustered samples 
are then used to estimate the parameters. It is a sub function used in the BMR.CTD function.
}
