% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter1.norm.R
\name{iter1.norm}
\alias{iter1.norm}
\title{iter1.norm}
\usage{
iter1.norm(x, para0, lambda)
}
\arguments{
\item{x}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{para0}{}

\item{lambda}{The level of penalty, default value is 0.}
}
\value{

}
\description{
iter1.norm
}
