% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter2.norm.R
\name{iter2.norm}
\alias{iter2.norm}
\title{iter2.norm}
\usage{
iter2.norm(x, para, beta, an, para0)
}
\arguments{
\item{x}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{para}{}

\item{beta}{}

\item{an}{}

\item{para0}{}
}
\value{

}
\description{
iter2.norm
}
