% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmm.norm.R
\name{maxmm.norm}
\alias{maxmm.norm}
\title{maxmm.norm}
\usage{
maxmm.norm(x, beta, an, para0, len, niter, tol)
}
\arguments{
\item{x}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{beta}{}

\item{an}{}

\item{para0}{}

\item{len}{The amount of initial values chosen for the EM-algorithm.}

\item{niter}{The least number of iterations for all initial values in the EM-algorithm.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}
}
\description{
maxmm.norm
}
