% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.bmm.R
\name{mom.beta}
\alias{mom.beta}
\title{mom.beta}
\usage{
mom.beta(data)
}
\arguments{
\item{data}{The input data.}
}
\description{
A sub function for MoM_Calculation, calculates the MoM for the beta distribution.
It is used in the MoM_Calculation function.
}
