% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.bmm.R
\name{mom.calculation}
\alias{mom.calculation}
\title{mom.calculation}
\usage{
mom.calculation(x, mix_porp, seed)
}
\arguments{
\item{x}{The input data, assumed to be from a beta mixture.}

\item{mix_porp}{A numeric vector of mixing proportions for each component.}

\item{seed}{An optional integer seed for reproducibility.}
}
\value{
A numeric vector containing concatenated alpha and beta estimates for each component.
}
\description{
A sub-function for \code{mom.bmm}, generates starting values for the method 
             of moments estimation of a beta mixture.
}
