% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.beta.R
\name{MoM_Beta}
\alias{MoM_Beta}
\title{MoM_Beta}
\usage{
MoM_Beta(data)
}
\arguments{
\item{data}{The input data.}
}
\description{
A sub function for MoM_Calculation, calculates the MoM for the beta distribution.
It is used in the MoM_Calculation function.
}
