% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty.norm.R
\name{penalty.norm.f}
\alias{penalty.norm.f}
\title{penalty.norm.f}
\usage{
penalty.norm.f(alpha, sigma, sigma0, an, lambda)
}
\arguments{
\item{alpha}{A vector of the mixture proportions of length m, m represents the order of the mixture}

\item{sigma}{A vector of the component standard deviations.}

\item{sigma0}{}

\item{an}{}

\item{lambda}{The level of penalty, default value is 0.}
}
\description{
penalty.norm.f
}
