% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.norm.R
\name{phi.norm}
\alias{phi.norm}
\title{phi.norm}
\usage{
phi.norm(x, m0, lambda, inival, len, niter, tol)
}
\arguments{
\item{x}{The data that can be either a vector or a matrix with the 1st column being the observed values}

\item{m0}{The order of the finite mixture model.}

\item{lambda}{The size of the penalized function of the mixing distribution.}

\item{inival}{The initial values chosen for the EM-algorithm.}

\item{len}{The number of initial values chosen for the EM-algorithm.}

\item{niter}{The smallest number of iterations required for all initial values in the EM-algorithm.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}
}
\value{

}
\description{
phi.norm
}
