% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi0.binom.R
\name{phi0.binom}
\alias{phi0.binom}
\title{phi0.binom}
\usage{
phi0.binom(x, size, m0, lambda, inival, len, niter, tol)
}
\arguments{
\item{x}{The data that can either be a vector or a matrix with the 1st column being the observed values.}

\item{size}{The size parameter for the Binomial mixture, and the 2nd column being the corresponding frequencies.}

\item{m0}{The order under the null hypothesis.}

\item{lambda}{The size of the penalty function of the mixing proportions.}

\item{inival}{The initial values chosen for the EM-algorithm.}

\item{len}{The number of initial values chosen for the EM-algorithm.}

\item{niter}{The smallest number of iterations required for all initial values in the EM-algorithm.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}
}
\description{
phi0.binom
}
