% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmix.binom.R
\name{plotmix.binom}
\alias{plotmix.binom}
\title{plotmix.binom}
\usage{
plotmix.binom(
  x,
  size,
  alpha,
  theta,
  sub.pmf = T,
  nclass = NULL,
  main = "",
  xlab = "Counts",
  ylab = "Prob"
)
}
\arguments{
\item{x}{The input data, either a vector or a matrix with two columns: count and freq.}

\item{size}{The number of trials.}

\item{alpha}{A vector of the mixing probabilities.}

\item{theta}{vector of probabilities of success of each component.}

\item{sub.pmf}{logical, draw subpopulation pmf's.}

\item{nclass}{The number of classes for a histogram.}

\item{main}{The main title of the output graph.}

\item{xlab}{The label name for the x-axis.}

\item{ylab}{The label name for the y-axis.}
}
\value{
It returns the histogram of the observations and the plot of the fitted probability mass function.
}
\description{
This function plots a histogram together with a suggested mixture pmf.
}
\examples{
alpha = c(.5, .1, .4)
theta = c(.1, .35, .8)
size = 25
x = rmix.binom(1000, size, alpha, theta)
plotmix.binom(x, size, alpha, theta, sub.pmf = TRUE, nclass=size, main="", xlab="Counts", ylab="Prob")
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
