% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmix.exp.R
\name{plotmix.exp}
\alias{plotmix.exp}
\title{plotmix.exp}
\usage{
plotmix.exp(
  alpha,
  mu,
  qq = 0.995,
  x = NULL,
  extra.height = 1.05,
  logObs = F,
  sub.pdf = T,
  h = 1.1,
  nclass = NULL,
  main = "",
  xlab = "Observations scaled",
  ylab = "Density",
  ylim = NULL
)
}
\arguments{
\item{alpha}{A vector of the mixing probabilities.}

\item{mu}{A vector of the subpopulation means.}

\item{qq}{The range of the plot, which is the qq th quantile.}

\item{x}{The data whose histogram is to be drawn from, can either be a vector or a matrix.}

\item{extra.height}{Additional height multiplier used to enlarge the plot vertically.}

\item{logObs}{logObs=T if data is in logarithm transformation form.}

\item{sub.pdf}{logical, draw subpopulation pdf's if True.}

\item{h}{A scale value used by ylim.}

\item{nclass}{The number of bars for the histogram if specified.}

\item{main}{The title for the histogram.}

\item{xlab}{The name of the x-axis label.}

\item{ylab}{The name of the y-axis label.}

\item{ylim}{The range of the y-axis, used only for the histogram if specified.}
}
\value{
It returns the histogram of observations and the plot of the fitted density
}
\description{
This function plots a histogram plus the corresponding fitted mixture pdf or the mixture pdf only, for a exponential mixture.
}
\examples{
n = 4000
mu = c(3, 9, 18)
alpha = c(.5, .3, .2)
x = rmix.exp(n, alpha, mu) 
plotmix.exp(alpha, mu, qq = 0.995, x) 
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
