% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmix.norm.bin.R
\name{plotmix.norm.bin}
\alias{plotmix.norm.bin}
\title{plotmix.norm.bin}
\usage{
plotmix.norm.bin(
  x = NULL,
  alpha,
  mu,
  sigma,
  k = 20,
  extra.height = 1.05,
  comp = T,
  hist.ind = T,
  main = "",
  xlab = "Observed values",
  ylab = "Density/Histogram"
)
}
\arguments{
\item{x}{The input data that can either be a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{alpha}{A vector of the mixing proportions.}

\item{mu}{A vector of the component means.}

\item{sigma}{A vector of the component standard deviations.}

\item{k}{The number of bars for the histogram.}

\item{extra.height}{Additional height multiplier used to enlarge the plot vertically.}

\item{comp}{A parameter for the component fitted density. 
comp=T means component fitted densities are drawn, and comp=F means no component fitted densities.}

\item{hist.ind}{It is true if there's data to be used to plot the histogram.}

\item{main}{The title of the graph.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}
}
\value{
It returns the histogram of observations and the plot of the fitted density
}
\description{
This function plots the mixture density, together with the histogram of the data
when the data generated from this density is given. It's similar to the plotmix.norm function, however 
this function manually draws the histogram and is more suitable for users that want to accurately control 
the number of bars that the histogram contains.
}
\examples{
n=5000
alpha = c(.2, .5, .3) 
mu = c(-1, 2, 6)
sigma = c(1, .8, 1.1)
x = rmix.norm(n, alpha, mu, sigma)
plotmix.norm.bin(x,alpha,mu,sigma)
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
