% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmix.pois.R
\name{plotmix.pois}
\alias{plotmix.pois}
\title{plotmix.pois}
\usage{
plotmix.pois(
  alpha,
  mu,
  x = NULL,
  x.max = NULL,
  sub.pmf = T,
  nclass = NULL,
  extra.height = 1.05,
  main = "",
  xlab = "Counts",
  ylab = "Prob"
)
}
\arguments{
\item{alpha}{A vector of the mixing proportions.}

\item{mu}{A vector of the subpopulation means.}

\item{x}{A vector of data whose histogram is to be drawn from.}

\item{x.max}{If x is null, 0:x.max is the range of pmf to be plotted.}

\item{sub.pmf}{logical, draw subpopulation pmf's.}

\item{nclass}{The number of bars for the histogram.}

\item{extra.height}{Additional height multiplier used to enlarge the plot vertically.}

\item{main}{The title for the histogram.}

\item{xlab}{The name of the x-axis label.}

\item{ylab}{The name of the y-axis label.}
}
\value{
It returns the histogram of the observations and the plot of the fitted probability mass function.
}
\description{
This function plots the data in a histogram with 
             the corresponding fitted mixture pmf
             or the mixture pmf only, for Poisson mixture. 
             The pmf is given as if it is a continuous curve.
}
\examples{
n = 2000
mu = c(2, 9, 11)
alpha = c(.2, .3, .5)
xx = rmix.pois(n, alpha, mu)
plotmix.pois(alpha,mu,xx,extra.height = 1.05)
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
