% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmix.exp.R
\name{pmix.exp}
\alias{pmix.exp}
\title{pmix.exp}
\usage{
pmix.exp(x, alpha, theta)
}
\arguments{
\item{x}{The value at which we computed the cdf.}

\item{alpha}{A vector of the mixing probabilities.}

\item{theta}{A vector of the subpopulation means.}
}
\description{
This function computes the cdf of the finite exponential mixture distribution
}
\examples{
n = 4000
theta = c(3, 9, 18)
alpha = c(.5, .3, .2)
x = rmix.exp(n, alpha, theta)
pmix.exp(x, alpha, theta)
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
