% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmix.norm.R
\name{pmix.norm}
\alias{pmix.norm}
\title{pmix.norm}
\usage{
pmix.norm(x, alpha, mu, sigma)
}
\arguments{
\item{x}{The cdf value of the Normal mixture at x.}

\item{alpha}{A vector of the mixing proportions.}

\item{mu}{A vector of the component means.}

\item{sigma}{A vector of the component standard deviations.}
}
\description{
pmix.norm
}
\examples{
n=3000
alpha = c(.2, .5, .3); mu = c(-2, 4, 5); 
sigma = c(1, .8, 1.1)*2
x = rmix.norm(n, alpha, mu, sigma)
pmix.norm(x,alpha,mu,sigma)
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
