% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.exp.R
\name{pmle.exp}
\alias{pmle.exp}
\title{pmle.exp}
\usage{
pmle.exp(
  x,
  m0,
  lambda = 1,
  init.val = NULL,
  n.init = 10,
  n.iter = 50,
  max.iter = 5000,
  tol = 1e-08,
  rformat = F
)
}
\arguments{
\item{x}{The input data that can either be a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{m0}{The order under the null hypothesis.}

\item{lambda}{The size of the penalized function of the mixing distribution, default value: lambda = 1.}

\item{init.val}{NULL or a 2 X m0 matrix with rows made of the mixing probabilities and component means.}

\item{n.init}{The initial values generated by the computer.}

\item{n.iter}{The number of EM iterations for each initial values. The one that gained the most in likelihood will be iterative further.}

\item{max.iter}{The maximum amount of iterations allowed. Set to 5000.}

\item{tol}{A tolerance value, EM concludes when the increment in p-likelihood is below tol.}

\item{rformat}{The format of the output, rformat=T means the format is determined by R.
rformat=F means the format is determined by our default setting. When the output is
larger than 0.001, it is determined by round(output,3); When the output is less than 0.001,
it is determined by signif(output,3).}
}
\value{
It returns the PMLE or MLE of the parameters with order = m0 (mixing proportions and component
parameters), log-likelihood value at the PMLE or MLE and the penalized log-likelihood value at the PMLE.
}
\description{
This function computes the penalized MLE, given iid observations under the finite exponential mixture model.
}
\examples{
n = 400
mu = c(3, 9, 18)
alpha = c(.5, .3, .2)
x = rmix.exp(n, alpha, mu) 
pmle.exp(x, 3)
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
