% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.exp.sub.R
\name{pmle.exp.sub}
\alias{pmle.exp.sub}
\title{pmle.exp.sub}
\usage{
pmle.exp.sub(xx, m0, lambda, init.val, n.init, n.iter, max.iter, tol)
}
\arguments{
\item{xx}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{m0}{The order under null hypothesis.}

\item{lambda}{The size of the penalized function of the mixing distribution, default value: lambda = 1.}

\item{init.val}{NULL or a 2 X m0 matrix with rows made of the mixing probabilities and component means.}

\item{n.init}{The initial values generated by the computer.}

\item{n.iter}{The number of EM iterations for each initial values. The one gained most in likelihood will be iterative further.}

\item{max.iter}{The maximum amount of iterations allowed. Set to 5000.}

\item{tol}{A tolerance value, EM concludes when the increment in p-likelihood is below tol.}
}
\description{
This function does the real work of computing the PMLE of the mixing distribution under the exponential mixture model. 
It is used in the pmle.exp function.
}
