% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.norm.R
\name{pmle.norm}
\alias{pmle.norm}
\title{pmle.norm}
\usage{
pmle.norm(
  x,
  m0,
  lambda = 1,
  an = NULL,
  init.val = NULL,
  n.init = 10,
  n.iter = 50,
  max.iter = 5000,
  tol = 1e-08,
  rformat = F
)
}
\arguments{
\item{x}{The data that can either be a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequencies.}

\item{m0}{The order of the finite mixture model, default value: m0 = 1.}

\item{lambda}{The size of the penalized function of the mixing distribution, default value: lambda = 1.}

\item{an}{A penalty on the variance(ChenTanZhangSinica2008). The recommended value is n^{-1/2}.}

\item{init.val}{NULL or a 3 X m0 matrix with rows made of mixing probability, component means, and variances.}

\item{n.init}{A computer generated n.init initials value.}

\item{n.iter}{The number of EM iterations for each initial value. The one that gained the most in likelihood will be iterative further.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm, default value: tol = 1e-6.}

\item{rformat}{The format of the output. If rformat=T, 
it means the output format is determined by R software. 
If rformat=F, it means the output format is determined by our default setting. 
When the output is larger than 0.001, it is determined by round(output,3); 
When the output is less than 0.001, it is determined by signif(output,3).
The default value of rformat is F.}
}
\value{
The PMLE or MLE of the parameters with order = m0 (mixing proportions, mixing means
         and mixing variances), log-likelihood value at the PMLE or MLE and the penalized log-likelihood
         value at the PMLE.
}
\description{
This function computes the PMLE or MLE of the parameters under a mixture of normals with unequal variance.
}
\examples{
 
data(grains)
pmle.norm(unlist(grains),2)
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
