% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmix.norm.R
\name{qmix.norm}
\alias{qmix.norm}
\title{qmix.norm}
\usage{
qmix.norm(pp, alpha, mu, sigma, tol = 1e-08)
}
\arguments{
\item{pp}{The level of the quantile to be computed.}

\item{alpha}{A vector of the mixing proportions.}

\item{mu}{A vector of the component means.}

\item{sigma}{A vector of the component standard deviations.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.#'}
}
\description{
This function computes the quantile of the univariate finite normal mixture one by one.
It is used in the tildeB22.norm0 function.
}
