% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmixnorm.R
\name{qmixnorm}
\alias{qmixnorm}
\title{qmixnorm}
\usage{
qmixnorm(alpha, theta, alp0)
}
\arguments{
\item{alpha}{A vector of the mixture probabilities.}

\item{theta}{A vector of means from each component.}

\item{alp0}{A given probability.}
}
\value{

}
\description{
This function computes the alp0-quantile of Normal mixture.
}
