% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual1.R
\docType{data}
\name{residual1}
\alias{residual1}
\title{residual1}
\format{
This data frame contains the 2 columns,
count: number of boys in family.
freq: number of families with corresponding.
}
\usage{
data(residual1)
}
\description{
This data set is from Lindsay and Roeder (1992). It includes the number of boys in families of size
8. The number of families is 53680. The data frame has 9 rows and 2 columns.
}
\references{
Lindsay, B. G. and Roeder, K. (1992). Residual diagnostics for mixture models. Journal of the
American Statistical Association,87(419), 785-794.
}
\keyword{datasets}
