% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmix.beta.R
\name{rmix.beta}
\alias{rmix.beta}
\title{rmix.beta}
\usage{
rmix.beta(n, mix_prop, alpha, beta)
}
\arguments{
\item{n}{The sample size.}

\item{mix_prop}{A vector of the mixing proportions.}

\item{alpha}{A vector of the component alphas.}

\item{beta}{A vector of the component betas.}
}
\value{
It returns a sample of size n from an m-component beta mixture.
}
\description{
This function generates the iid samples from the finite beta mixture
}
\examples{
n = 1000
mix_prop = c(.5, .3, .2)
alpha = c(1,20,3)
beta = c(10,0.5,2)
rmix.beta(n, mix_prop, alpha,beta)
}
