% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmix.binom.R
\name{rmix.binom}
\alias{rmix.binom}
\title{rmix.binom}
\usage{
rmix.binom(n, size, alpha, theta)
}
\arguments{
\item{n}{The sample size.}

\item{size}{The number of trials.}

\item{alpha}{A vector of the mixing proportions.}

\item{theta}{A vector of the probability of success of each component.}
}
\value{
It returns the sample of size n from an m-component binomial mixture.
}
\description{
This function generates a random sample from the finite binomial mixture distributions.
}
\examples{
 alpha = c(.5, .1, .4)
theta = c(.1, .35, .8)
size = 25
x = rmix.binom(1000, size, alpha, theta)
mean(x)
var(x)
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
