% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmix.norm.R
\name{rmix.norm}
\alias{rmix.norm}
\title{rmix.norm}
\usage{
rmix.norm(n, alpha, mu, sigma = rep(1, length(alpha)))
}
\arguments{
\item{n}{The sample size.}

\item{alpha}{A vector of the mixture proportions of length m, also the order of the mixture.}

\item{mu}{A vector of the component distributions' mean.}

\item{sigma}{A vector of the component distributions' standard deviation, default value: sigma =
rep(1,length(alpha)).}
}
\value{
It returns a samples of size n from an m-component normal mixture.
}
\description{
This function generates a random sample from a mixture of normal.
}
\examples{
x <- rmix.norm(200,c(0.3,0.7),c(-1,2),c(1,2))
hist(x)
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
