% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmix.pois.R
\name{rmix.pois}
\alias{rmix.pois}
\title{rmix.pois}
\usage{
rmix.pois(n, alpha, theta)
}
\arguments{
\item{n}{The sample size.}

\item{alpha}{A vector of the mixing proportions.}

\item{theta}{A vector of the subpopulation means.}
}
\value{
It returns a sample of size n from an m-component Poisson mixture.
}
\description{
This function generates a random sample from a finite Poisson mixture
}
\examples{
n = 3000
mu = c(3, 9, 15)
alpha = c(.2, .3, .5)
xx = rmix.pois(n, alpha, mu)
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
