% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar.beta.mix.R
\name{sar.beta.mix.sub}
\alias{sar.beta.mix.sub}
\title{sar.beta.mix.sub}
\usage{
sar.beta.mix.sub(x, w, epsilon)
}
\arguments{
\item{x}{A numeric vector of observations in (0, 1).}

\item{w}{A matrix of component-wise weights.}

\item{epsilon}{A regularization parameter controlling the smoothing in the E-step. Default is \code{1}.}
}
\value{
A numeric vector of length 3m: concatenated estimates of
\item{pi}{Mixing proportions.}
\item{alpha}{Component-wise alpha parameters.}
\item{beta}{Component-wise beta parameters.}
}
\description{
A sub-function of the main function \code{sar.beta.mix}.
}
