% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar.beta.mix.R
\name{sar.calculation}
\alias{sar.calculation}
\title{sar.calculation}
\usage{
sar.calculation(x, cluster_assignments, m0)
}
\arguments{
\item{x}{A numeric vector of data in (0, 1).}

\item{cluster_assignments}{Integer vector assigning each observation to one of the m0 components.}

\item{m0}{Number of mixture components.}
}
\value{
A numeric vector: \code{c(alpha_1, ..., alpha_m0, beta_1, ..., beta_m0)}.
}
\description{
Computes unweighted SAR estimates of beta parameters for each mixture component, 
             based on initial hard assignments (e.g., from K-means clustering).
}
