% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb2.exp.R
\name{tb2.exp}
\alias{tb2.exp}
\title{tb2.exp}
\usage{
tb2.exp(alpha, theta)
}
\arguments{
\item{alpha}{A vector of the mixing probabilities.}

\item{theta}{A vector of the means of each component.}
}
\description{
This function computes \tilde B_{22} under the Exponential mixture.
}
