% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb2.norm.R
\name{tb2.norm}
\alias{tb2.norm}
\title{tb2.norm}
\usage{
tb2.norm(alpha, theta)
}
\arguments{
\item{alpha}{A vector of the mixture probabilities.}

\item{theta}{A vector of the means of each component.}
}
\value{

}
\description{
This function computes \tilde B_{22} under the normal mixture.
\tilde B_{22} has been standardized into a correlation matrix.
To calculate \tilde B_{22}, (-\infty,\infty) has been divided into N=10000 subintervals with the probability in each subinterval being 1/N.
In practice, you may change N to be a large value to see its effect.
}
