% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tildeB22.exp.R
\name{tildeB22.exp}
\alias{tildeB22.exp}
\title{tildeB22.exp}
\usage{
tildeB22.exp(alpha, mu)
}
\arguments{
\item{alpha}{A vector of the mixing probabilities.}

\item{mu}{A vector of the means of each component.}
}
\description{
This function computes tilde{B}_{22} under the Exponential mixture.
It is used in the emtest.exp function.
}
