% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMR.CTD.R
\name{beta.barycenter}
\alias{beta.barycenter}
\title{beta.barycenter}
\usage{
beta.barycenter(
  assignments,
  M,
  cluster_weights,
  cluster_alphas,
  cluster_betas,
  divergence = "Hellinger"
)
}
\arguments{
\item{assignments}{An integer vector indicating the cluster assignment of each original beta component.}

\item{M}{The number of target components in the reduced mixture.}

\item{cluster_weights}{A numeric vector of weights for each original beta component.}

\item{cluster_alphas}{A numeric vector of \eqn{\alpha} parameters for each original beta component.}

\item{cluster_betas}{A numeric vector of \eqn{\beta} parameters for each original beta component.}

\item{divergence}{Character string, either "Hellinger" (default) or "SKL"}
}
\value{
A matrix with \code{M} rows and 2 columns, where each row contains the estimated 
        \code{(alpha, beta)} parameters of the corresponding cluster's barycenter.
}
\description{
Computes the barycenter of a collection of beta distributions 
with respect to a hybrid divergence—a convex combination of the squared Hellinger distance 
and the symmetrized KL divergence. This function is a subroutine used in \code{BMR.CTD}.
}
