% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMR.CTD.R
\name{beta.barycenter}
\alias{beta.barycenter}
\title{beta.barycenter}
\usage{
beta.barycenter(cluster_weights, cluster_alphas, cluster_betas)
}
\arguments{
\item{cluster_weights}{A numeric vector of weights associated with each beta distribution.}

\item{cluster_alphas}{A numeric vector of shape parameters \eqn{\alpha} for each beta distribution.}

\item{cluster_betas}{A numeric vector of shape parameters \eqn{\beta} for each beta distribution.}
}
\value{
A numeric vector of length two containing the estimated \code{(alpha, beta)} parameters 
        of the barycenter.
}
\description{
Computes the barycenter of a collection of beta distributions 
with respect to the Hellinger distance, using a gradient-based optimization method.
It is a sub function used in the BMR.CTD function.
}
