% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.ident.beta.R
\name{check.comp}
\alias{check.comp}
\title{check.comp}
\usage{
check.comp(indices, mix_porp, alpha, beta, tol = 1e-06)
}
\arguments{
\item{indices}{The index numbers of the original input parameters.}

\item{mix_porp}{A vector of mixing proportions.}

\item{alpha}{A vector of subpopulation alphas.}

\item{beta}{A vector of subpopulation betas.}

\item{tol}{The tolerance value for checking the difference between actual and expected weights.}
}
\description{
The sub function that checks whether a group of components can be merged.
}
