% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix.binom.R
\name{dmix.binom}
\alias{dmix.binom}
\title{dmix.binom}
\usage{
dmix.binom(x, size, alpha, theta)
}
\description{
This function computes the probability mass function points for vector x.
It is used in the emstat.binom function.
}
