% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix.exp.R
\name{dmix.exp}
\alias{dmix.exp}
\title{dmix.exp}
\usage{
dmix.exp(x, alpha, mu, logObs = F)
}
\description{
This function computes the pdf of the finite exponential mixture distribution.
It is used in various functions within the Exponential folder.
}
