% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix.exp.R
\name{dmix.exp}
\alias{dmix.exp}
\title{dmix.exp}
\usage{
dmix.exp(x, alpha, mu, logObs = F)
}
\arguments{
\item{x}{The value at which we computed the cdf.}

\item{alpha}{A vector of the mixing probabilities.}

\item{mu}{A vector of the subpopulation means.}

\item{logObs}{logObs=T if the data is in logarithm}
}
\description{
This function computes the pdf of the finite exponential mixture distribution.
}
\examples{
alpha = c(.3, .7)
mu = c(1.8, 0.9)
n = 2000
x = rmix.exp(n, alpha = alpha, theta = mu)
dmix.exp(x,alpha,mu)
}
