% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix.norm.R
\name{dmix.norm}
\alias{dmix.norm}
\title{dmix.norm}
\usage{
dmix.norm(xx, alpha, mu, sigma)
}
\description{
A function that computes the density of the univariate normal mixture.
It is used in various functions within the Gaussian folder.
}
