% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix.norm.R
\name{dmix.norm}
\alias{dmix.norm}
\title{dmix.norm}
\usage{
dmix.norm(xx, alpha, mu, sigma)
}
\arguments{
\item{alpha}{A vector of the mixing proportions.}

\item{mu}{A vector of the component means.}

\item{sigma}{A vector of the component standard deviations.}

\item{x}{uni-variate mixture density at x.}

\item{tol}{A fixed value at 1e-8}
}
\description{
A function that computes the density of the univariate normal mixture
}
