% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix.norm.R
\name{dmix.norm}
\alias{dmix.norm}
\title{dmix.norm}
\usage{
dmix.norm(xx, alpha, mu, sigma)
}
\arguments{
\item{xx}{The uni-variate mixture density at x.}

\item{alpha}{A vector of the mixing proportions.}

\item{mu}{A vector of the component means.}

\item{sigma}{A vector of the component standard deviations.}
}
\description{
A function that computes the density of the univariate normal mixture.
}
\note{
In order to plot the mixture density together with its subpopulation density, one can use the plotmix.norm and plotmix.norm.bin functions.
}
\examples{
mu = c(8, 9)
alpha = c(.2, .8)
sigma = rep(1,length(alpha))
n = 1000
x = rmix.norm(n,alpha,mu)
dmix.norm(1:10,alpha,mu,sigma)

}
