% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix.pois.R
\name{dmix.pois}
\alias{dmix.pois}
\title{dmix.pois}
\usage{
dmix.pois(xx, alpha, theta)
}
\description{
This function computes the probability mass function of the Poisson mixture.
It is used in the emstat.pois function.
}
