% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\docType{data}
\name{earthquake}
\alias{earthquake}
\title{earthquake}
\format{
This data frame contains one column:
number: number of major earthquakes in a year.
}
\usage{
data(earthquake)
}
\description{
This data set contains the number of major earthquakes (magnitude 7 or greater) in the world from
1900 to 2006. The data are available in Table 1.1 of Zucchini & MacDonald (2009). Zucchini &
MacDonald (2009) suggested using a Poisson mixture model to fit the data. The data frame has 107
rows and 1 column.
}
\references{
Zucchini W, MacDonald IL (2009). Hidden Markov models for time series: an introduction using
R. CRC Press.
}
\keyword{datasets}
