% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emiter.norm.R
\name{emiter.norm}
\alias{emiter.norm}
\title{emiter.norm}
\usage{
emiter.norm(x, para, beta, pens, para0, k)
}
\arguments{
\item{x}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{para}{}

\item{beta}{}

\item{pens}{A 2-dimensions vector being the size of the penalized functions for 
the mixing propotion and the variance.}

\item{para0}{}

\item{k}{The amount of EM iterations in order to obtain the EM-test statistic.}
}
\description{
emiter.norm
}
