% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emstat.binom.R
\name{emstat.binom}
\alias{emstat.binom}
\title{emstat.binom}
\usage{
emstat.binom(
  count,
  freq,
  outnull,
  size,
  m0,
  CC,
  n.init,
  max.iter,
  n.iter,
  tol,
  k
)
}
\description{
This function computes the EM-test statistics for a binomial mixture.
It is used in the emtest.binom function.
}
