% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emstat.exp.R
\name{emstat.exp}
\alias{emstat.exp}
\title{emstat.exp}
\usage{
emstat.exp(x, outnull, C, len, niter, tol, k)
}
\arguments{
\item{x}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{outnull}{The output from the phi0.pois function.}

\item{C}{A optional tuning parameter for the EM-test procedure.}

\item{len}{The number of initial values chosen for the EM-algorithm.}

\item{niter}{The least amount of iterations for all initial values in the EM-algorithm.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}

\item{k}{The number of EM iterations needed to obtain the EM-test statistic.}
}
\description{
This function computes the EM-test statistics for the exponential mixture.
}
