% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emstat.exp.R
\name{emstat.exp}
\alias{emstat.exp}
\title{emstat.exp}
\usage{
emstat.exp(x, outnull, CC, n.init, n.iter, tol, k, max.iter)
}
\arguments{
\item{x}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{outnull}{The output from the phi0.pois function.}

\item{CC}{A optional tuning parameter for the EM-test procedure.}

\item{n.init}{The number of initial values chosen for the EM-algorithm.}

\item{n.iter}{The initial number of EM iterations.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}

\item{k}{The number of EM iterations needed to obtain the EM-test statistic.}

\item{max.iter}{The maximum number of iterations.}
}
\description{
This function computes the EM-test statistics for the exponential mixture.
}
