% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emstat.exp.R
\name{emstat.exp}
\alias{emstat.exp}
\title{emstat.exp}
\usage{
emstat.exp(x, outnull, CC, n.init, n.iter, tol, k, max.iter)
}
\description{
This function computes the EM-test statistics for the exponential mixture.
It is used in the emtest.exp function.
}
