% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emstat.norm0.R
\name{emstat.norm0}
\alias{emstat.norm0}
\title{emstat.norm0}
\usage{
emstat.norm0(xx, mle0, ln0, m0, CC, n.init, n.iter, tol, k = 3)
}
\description{
The function computes the EM-test statistics for univariate normal mixture
with known variance. It is used in the emtest.norm0 function.
}
