% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMstat.pois.R
\name{EMstat.pois}
\alias{EMstat.pois}
\title{EMstat.pois}
\usage{
EMstat.pois(
  count,
  freq,
  alpha0,
  theta0,
  m0,
  ln0,
  CC,
  n.init,
  n.iter,
  tol,
  k,
  max.iter
)
}
\description{
This function computes the EM-test statistics for Poisson mixture. 
It is used in the emtest.pois function.
}
