% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMstat.pois.R
\name{EMstat.pois}
\alias{EMstat.pois}
\title{EMstat.pois}
\usage{
EMstat.pois(
  count,
  freq,
  alpha0,
  theta0,
  m0,
  ln0,
  CC,
  n.init,
  n.iter,
  tol,
  k,
  max.iter
)
}
\arguments{
\item{count}{The observed values for the Poisson count data.}

\item{freq}{The corresponding frequency of the above counts.}

\item{alpha0}{The mixing proportion fitted under the null model.}

\item{theta0}{The subpopulation means fitted under the null model.}

\item{m0}{The order of the mixture to be fitted.}

\item{ln0}{The log likelihood of the best null model obtained from previous function.}

\item{CC}{The optional tuning parameter for EM-test procedure.}

\item{n.init}{The number of initial values chosen for the EM-algorithm.}

\item{n.iter}{Least amount of iterations for all initial values in the EM-algorithm.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}

\item{k}{The number of EM iterations required in order to obtain EM-test statistic.}

\item{max.iter}{Maximum amount of iterations.}
}
\description{
This function computes the EM-test statistics for Poisson mixture.
}
