% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emtest.binom.R
\name{emtest.binom}
\alias{emtest.binom}
\title{emtest.binom}
\usage{
emtest.binom(
  x,
  size,
  m0,
  CC = NULL,
  lambda = 0,
  init.val = NULL,
  n.init = 10,
  n.iter = 50,
  max.iter = 2000,
  tol = 1e-06,
  k = 3,
  rformat = FALSE
)
}
\arguments{
\item{x}{The input data, either a vector or a matrix columns with two columns: count and freq.}

\item{size}{The number of trials of the binomial.}

\item{m0}{The order under the null hypothesis.}

\item{CC}{A optional tuning parameter for the EM-test procedure.}

\item{lambda}{A tuning for the fitting null model.}

\item{init.val}{The initial values for the PMLE under the null model.}

\item{n.init}{The number of initial values chosen for the EM-algorithm.}

\item{n.iter}{The least amount of iterations for all initial values.}

\item{max.iter}{The maximum amount of iterations allowed.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}

\item{k}{The number of EM iterations in order to obtain the EM-test statistic.}

\item{rformat}{The format of the output. See rforma().}
}
\description{
This function computes the EM-test statistic and the p-value for H0: m=m0.
}
\examples{
alpha = c(.5, .1, .4)
theta = c(.1, .35, .8)
size = 25
x = rmix.binom(1000, size, alpha, theta)
y = as.matrix(table(x))
count = as.numeric(rownames(y))
freq = y[,1]
emtest.binom(x, size, m0=2, CC=NULL, lambda = 0, init.val=NULL,
n.init=10, n.iter=50, max.iter=200, tol=1e-6, k=3, rformat=FALSE)
}
