% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMtest.maxmm.norm.R
\name{EMtest.maxmm.norm}
\alias{EMtest.maxmm.norm}
\title{EMtest.maxmm.norm}
\usage{
EMtest.maxmm.norm(xx, beta.i, m0, an, para0, n.init, n.iter, tol)
}
\arguments{
\item{xx}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{beta.i}{The fixed mixing proportions.}

\item{m0}{The order of the finite normal mixture model.}

\item{an}{A penalty on the variance(ChenTanZhangSinica2008). The recommended value is n^{-1/2}.}

\item{para0}{The subpopulations parameter.}

\item{n.init}{A computer generated n.init initials value.}

\item{n.iter}{The number of EM iterations for each initial values. The one gained the most in likelihood will be iterative further.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}
}
\description{
This function computes the mixture's PMLE with specific structure. 
It is used in the emtest.norm function.
}
