% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMtest.maxmm.norm.R
\name{EMtest.maxmm.norm}
\alias{EMtest.maxmm.norm}
\title{EMtest.maxmm.norm}
\usage{
EMtest.maxmm.norm(xx, beta.i, m0, an, para0, n.init, n.iter, tol)
}
\description{
This function computes the mixture's PMLE with specific structure. 
It is used in the emtest.norm function.
}
