% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMtest.maxmm.norm0.R
\name{EMtest.maxmm.norm0}
\alias{EMtest.maxmm.norm0}
\title{EMtest.maxmm.norm0}
\usage{
EMtest.maxmm.norm0(xx, bbeta, mle0, m0, n.init, n.iter, tol)
}
\arguments{
\item{xx}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency. It is used in the emstat.norm0 function.}
}
\description{
EMtest.maxmm.norm0
}
