% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMtest.maxmm.norm0.R
\name{EMtest.maxmm.norm0}
\alias{EMtest.maxmm.norm0}
\title{EMtest.maxmm.norm0}
\usage{
EMtest.maxmm.norm0(xx, bbeta, mle0, m0, n.init, n.iter, tol)
}
\arguments{
\item{xx}{}

\item{bbeta}{The fixed mixing proportions.}

\item{mle0}{The subpopulation means fitted under the null hypothesis}

\item{m0}{}

\item{n.init}{The number of initial values used.}

\item{n.iter}{The number of iterations from each initial.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}
}
\description{
EMtest.maxmm.norm0
}
