% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMtest.maxmm.pois.R
\name{EMtest.maxmm.pois}
\alias{EMtest.maxmm.pois}
\title{EMtest.maxmm.pois}
\usage{
EMtest.maxmm.pois(
  count,
  freq,
  beta.i,
  alpha0,
  theta0,
  m0,
  n.init,
  n.iter,
  tol,
  max.iter
)
}
\arguments{
\item{count}{The observed values for the Poisson count data.}

\item{freq}{The corresponding frequency of the above counts.}

\item{beta.i}{The fixed mixing proportions.}

\item{alpha0}{The mixing proportion fitted under the null model.}

\item{theta0}{The subpopulation means fitted under the null model.}

\item{m0}{The order of the mixture to be fitted.}

\item{n.init}{The number of initial values chosen for the EM-algorithm.}

\item{n.iter}{Least amount of iterations for all initial values in the EM-algorithm.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}

\item{max.iter}{Maximum amount of iterations.}
}
\description{
This function computes the PMLE of parameters under the alternative model 
             for EM-test given a beta.i value. It is used in the EMstat.pois function.
}
