% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMtest.maxmm.pois.R
\name{EMtest.maxmm.pois}
\alias{EMtest.maxmm.pois}
\title{EMtest.maxmm.pois}
\usage{
EMtest.maxmm.pois(
  count,
  freq,
  beta.i,
  alpha0,
  theta0,
  m0,
  n.init,
  n.iter,
  tol,
  max.iter
)
}
\description{
This function computes the PMLE of parameters under the alternative model 
             for EM-test given a beta.i value.
             It is used in the EMstat.pois function.
}
