% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMtest.norm.iter.R
\name{EMtest.norm.iter}
\alias{EMtest.norm.iter}
\title{EMtest.norm.iter}
\usage{
EMtest.norm.iter(xx, nn, m0, para, sigma0, beta.i, pen.size, k)
}
\arguments{
\item{xx}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{m0}{The order of the finite normal mixture model.}

\item{beta.i}{The fixed mixing proportions.}

\item{an}{}

\item{para0}{The subpopulations parameter.}

\item{n.init}{A computer generated n.init initials value.}

\item{n.iter}{The number of EM iterations for each initial values. The one gained the most in likelihood will be iterative further.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}
}
\description{
This function computes the EM-algorithm in the emtest.norm function.
}
