% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMtest.norm.iter.R
\name{EMtest.norm.iter}
\alias{EMtest.norm.iter}
\title{EMtest.norm.iter}
\usage{
EMtest.norm.iter(xx, nn, m0, para, sigma0, beta.i, pen.size, k)
}
\description{
This function computes the EM-algorithm in the emtest.norm function.
}
