% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMtest.norm.iter.R
\name{EMtest.norm.iter}
\alias{EMtest.norm.iter}
\title{EMtest.norm.iter}
\usage{
EMtest.norm.iter(xx, nn, m0, para, sigma0, beta.i, pen.size, k)
}
\arguments{
\item{xx}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{nn}{}

\item{m0}{The order of the finite mixture model.}

\item{para}{}

\item{sigma0}{}

\item{beta.i}{}

\item{pen.size}{}

\item{k}{}
}
\description{
EMtest.norm.iter
}
