% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMtest.norm.penalty.R
\name{EMtest.norm.penalty}
\alias{EMtest.norm.penalty}
\title{EMtest.norm.penalty}
\usage{
EMtest.norm.penalty(m0, para0, nn)
}
\arguments{
\item{m0}{The order of the finite normal mixture model.}

\item{para0}{The subpopulations parameter.}

\item{nn}{The lenght of the input data.}
}
\description{
This function produces the recommended size of penalty of EMtest under the Gaussian mixture.
It is used in the emtest.norm function.
}
