% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMtest.norm.penalty.R
\name{EMtest.norm.penalty}
\alias{EMtest.norm.penalty}
\title{EMtest.norm.penalty}
\usage{
EMtest.norm.penalty(m0, para, nn)
}
\arguments{
\item{m0}{The order}

\item{para}{The list of mixing proportion and its mean and var}

\item{nn}{The sample size}
}
\description{
EMtest.norm.penalty
}
