% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emtest.norm.R
\name{emtest.norm}
\alias{emtest.norm}
\title{emtest.norm}
\usage{
emtest.norm(
  x,
  m0 = 1,
  pens = NULL,
  inival = NULL,
  len = 10,
  niter = 50,
  tol = 1e-06,
  k = 3,
  rformat = FALSE
)
}
\arguments{
\item{x}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{m0}{The order of the finite normal mixture model under the null hypothesis.}

\item{pens}{A 2-dimensions vector being the size of the penalized functions for 
the mixing propotion and the variance.}

\item{inival}{The initial values chosen for the EM-algorithm to compute the PMLE under the null model}

\item{len}{The amount of initial values chosen for the EM-algorithm.}

\item{niter}{The least number of iterations for all initial values in the EM-algorithm.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}

\item{k}{The amount of EM iterations in order to obtain the EM-test statistic.}

\item{rformat}{F means the format of output is determined by our default setting. When the output is
larger than 0.001, it is determined by round(output,3); When the output is less than 0.001,
it is determined by signif(output,3).}
}
\description{
emtest.norm
}
