% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emtest.norm0.thm3.R
\name{emtest.norm0.thm3}
\alias{emtest.norm0.thm3}
\title{emtest.norm0.thm3}
\usage{
emtest.norm0.thm3(tb, N = 10000, tol = 1e-08)
}
\arguments{
\item{N}{The number of repetitions.}

\item{tol}{A value below which is judged as 0.}
}
\description{
This function computes a_h values in Theorem 3 of Li and Chen (JASA2009) by simulation
}
\note{
Need to first install the R package quadprog.
}
