% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emtest.thm3.R
\name{emtest.norm0.thm3}
\alias{emtest.norm0.thm3}
\title{emtest.thm3}
\usage{
emtest.norm0.thm3(tb, N = 10000, tol = 1e-08)
}
\description{
This function computes a_h as in Theorem 3 of Li and Chen (JASA2009) by MC.
It is used in the emtest.exp function.
}
