% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emtest.pois.R
\name{emtest.pois}
\alias{emtest.pois}
\title{emtest.pois}
\usage{
emtest.pois(
  x,
  m0 = 1,
  CC = NULL,
  init.val = NULL,
  n.init = 10,
  n.iter = 50,
  tol = 1e-06,
  k = 3,
  rformat = F,
  max.iter = 5000
)
}
\arguments{
\item{x}{Data that can be either a vector or a matrix of observations. 
1st column: counts 
2nd column: frequencies.}

\item{CC}{The optional tuning parameter for EM-test procedure.}

\item{init.val}{A user provided initial values for the EM-algorithm to 
compute the PMLE under the null model.}

\item{n.init}{The number of initial values chosen for the EM-algorithm.}

\item{n.iter}{Least amount of iterations for all initial values in the EM-algorithm.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}

\item{k}{The number of EM iterations required in order to obtain EM-test statistic.}

\item{rformat}{A specific format, please see rousignif.R function.}

\item{max.iter}{Maximum amount of iterations.}
}
\description{
This function computes the EM-test statistic and p-value for the null hypothesis that the order is m = m0.
}
