% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmm.binom.R
\name{maxmm.binom}
\alias{maxmm.binom}
\title{maxmm.binom}
\usage{
maxmm.binom(
  count,
  freq,
  size,
  m0,
  beta.j,
  theta0,
  n.init,
  max.iter,
  n.iter,
  tol
)
}
\description{
This function is meant for em-test. It gets the PMLE under a restricted alternative given a beta.j value.
It is used in the emtest.binom function.
}
