% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmm.exp.R
\name{maxmm.exp}
\alias{maxmm.exp}
\title{maxmm.exp}
\usage{
maxmm.exp(x, bbeta, theta0, n.init, n.iter, tol, max.iter)
}
\description{
This function computes the PMLE under the alternative model given a beta value. 
It is used in the emstat.exp function.
}
