% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmm.exp.R
\name{maxmm.exp}
\alias{maxmm.exp}
\title{maxmm.exp}
\usage{
maxmm.exp(x, beta, theta0, len, niter, tol)
}
\arguments{
\item{x}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{beta}{}

\item{theta0}{}

\item{len}{The number of initial values chosen for the EM-algorithm.}

\item{niter}{The least amount of iterations for all initial values in the EM-algorithm.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm.}
}
\description{
This function computes the PMLE of parameters under the alternative model given a beta value.
}
