% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.bmm.R
\name{mom.beta}
\alias{mom.beta}
\title{mom.beta}
\usage{
mom.beta(data)
}
\arguments{
\item{data}{A numeric vector assumed to be sampled from a beta distribution.}
}
\value{
A list with components:
\item{alpha}{The estimated alpha parameter.}
\item{beta}{The estimated beta parameter.}
}
\description{
A sub-function for \code{mom.calculation}, estimates the parameters of a single beta 
             distribution using the method of moments.
}
