% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.bmm.R
\name{mom.bmm}
\alias{mom.bmm}
\title{mom.bmm}
\usage{
mom.bmm(x, m0, maxit)
}
\arguments{
\item{x}{The input data, assumed to follow a beta mixture distribution.}

\item{m0}{The number of components in the mixture.}

\item{maxit}{Maximum number of iterations allowed in the nonlinear solver.}
}
\value{
A matrix of estimated initial parameters with valid solutions.
}
\description{
A sub-function for \code{pmle.beta}, performs the method of moments (MoM) estimation 
             for the beta mixture model.
}
