% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.bmm.R
\name{mom.calculation}
\alias{mom.calculation}
\title{mom.calculation}
\usage{
mom.calculation(x, cluster_assignments, m0)
}
\arguments{
\item{x}{The input data, assumed to be from a beta mixture.}

\item{cluster_assignments}{A vector of component identity that identifies which subpopulation that this observed value belongs to.}

\item{m0}{The number of components in the mixture.}
}
\value{
A numeric vector containing concatenated alpha and beta estimates for each component.
}
\description{
A sub-function for \code{mom.bmm}, generates starting values for the method 
             of moments estimation of a beta mixture.
}
