% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.bmm.R
\name{mom.calculation}
\alias{mom.calculation}
\title{mom.calculation}
\usage{
mom.calculation(x, mix_porp, seed)
}
\arguments{
\item{x}{The input data.}

\item{mix_porp}{The mixing proportion for each component/subpopulation.}

\item{seed}{For reproducible results.}
}
\description{
A sub function for mom.bmm, generates starting points for the mom.bmm function.
It is used in the mom.bmm function.
}
