% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.beta.sub.R
\name{Pen.M.Step}
\alias{Pen.M.Step}
\title{Pen.M.Step}
\usage{
Pen.M.Step(x, ww, mix_prop, theta, an)
}
\arguments{
\item{x}{A numeric vector of observed values assumed to follow a beta mixture distribution.}

\item{ww}{A matrix of posterior probabilities with dimensions \code{n x m0}, representing 
the probability that each observation belongs to each mixture component.}

\item{mix_prop}{A numeric vector of mixing proportions for each component.}

\item{theta}{A numeric vector of current estimates for \eqn{\alpha} and \eqn{\beta} 
parameters for all components.}

\item{an}{A penalty control parameter, typically \eqn{n^{-3/2}}, that influences the strength 
of the regularization.}
}
\value{
A numeric vector of updated parameter estimates for \eqn{\alpha} and \eqn{\beta}.
}
\description{
A sub-function for \code{pmle.beta.sub}, performs the M-step in the 
             EM algorithm for beta mixture.
}
