% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.beta.sub.R
\name{Pen.M.Step}
\alias{Pen.M.Step}
\title{Pen_M_Step}
\usage{
Pen.M.Step(x, ww, mix_porp, theta, an)
}
\arguments{
\item{x}{A vector of the observed values.}

\item{ww}{The probability that data point i belongs to component j (matrix with dimensions n x m0).}

\item{mix_porp}{The mixing proportions for each component.}

\item{theta}{A vector of estimated alpha and beta for every component.}

\item{an}{A size control parameter that controls the severity of the penalty.}
}
\description{
A sub function for pmle.beta.sub, does the actual work of the M step in the EM algorithm.
It is used in the pmle.beta.sub function.
}
