% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.beta.sub.R
\name{Pen_M_Step}
\alias{Pen_M_Step}
\title{Pen_M_Step}
\usage{
Pen_M_Step(x, ww, mix_porp, theta, an)
}
\description{
A sub function for pmle.beta.sub, does the actual work of the M step in the EM algorithm.
It is used in the pmle.beta.sub function.
}
