% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmix.beta.R
\name{plotmix.beta}
\alias{plotmix.beta}
\title{plotmix.beta}
\usage{
plotmix.beta(
  x,
  xx.grid = NULL,
  porp,
  alpha,
  beta,
  m0,
  k = 20,
  extra.height = 1.05,
  comp = T,
  main = "",
  xlab = "Observed values",
  ylab = "Density/Histogram"
)
}
\arguments{
\item{x}{The input data which is a vector.}

\item{xx.grid}{The grid of the histogram.}

\item{porp}{A vector of the mixing proportions.}

\item{alpha}{A vector of the component alphas.}

\item{beta}{A vector of the component betas.}

\item{m0}{The order of the finite mixture model.}

\item{k}{The number of bars for the histogram.}

\item{extra.height}{Additional height multiplier used to enlarge the plot vertically.}

\item{comp}{A parameter for the component fitted density. 
comp=T means component fitted densities are drawn, and comp=F means no component fitted densities.}

\item{main}{The title of the graph.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}
}
\value{
It returns the histogram of observations and the plot of the fitted density
}
\description{
This function plots the mixture density, together with the histogram of the data
when the data generated from this density is given.
}
\examples{
n=2000
porp = c(.2, .5, .3)
alpha = c(3, 2, 5)
beta = c(1, .5, 1.1)
x = rmix.beta(n, porp, alpha, beta)
plotmix.beta(x, xx.grid=NULL, porp, alpha, beta, m0=3,
main="", xlab="Observed values", ylab="Density/Histogram")
}
