% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmix.exp.R
\name{plotmix.exp}
\alias{plotmix.exp}
\title{plotmix.exp}
\usage{
plotmix.exp(
  alpha,
  mu,
  qq = 0.995,
  x = NULL,
  logObs = F,
  sub.pdf = T,
  h = 1.1,
  nclass = NULL,
  main = "",
  xlab = "Observations scaled",
  ylab = "Density",
  ylim = NULL
)
}
\arguments{
\item{alpha}{A vector of the mixing probabilities.}

\item{mu}{A vector of the subpopulation means.}

\item{qq}{The range of the plot, which is the qq th quantile.}

\item{x}{The data whose histogram is to be drawn, can either be a vector or a matrix.}

\item{logObs}{logObs=T if data is in logarithm}

\item{sub.pdf}{logical, draw subpopulation pdf's if True.}

\item{h}{A scale value used by ylim.}

\item{nclass}{The number of bars for the histogram if specified.}

\item{main}{The title for the histogram.}

\item{xlab}{The name of the x-axis label.}

\item{ylab}{The name of the y-axis label.}

\item{ylim}{Used only for the histogram if specified.}
}
\description{
This function plots a histogram plus the corresponding fitted mixture pdf or the mixture pdf only, for exponential mixture.
}
