% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmix.norm0.R
\name{plotmix.norm0}
\alias{plotmix.norm0}
\title{plotmix.norm0}
\usage{
plotmix.norm0(
  x,
  var,
  theta,
  hist = 1,
  comp = TRUE,
  k = 20,
  h = 1,
  main = "",
  xlab = "Observations",
  ylab = ""
)
}
\arguments{
\item{x}{The input data that can be either a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequency.}

\item{var}{The known component variance.}

\item{theta}{The output of pmle.norm, emtest.norm or the parameter values. This includes the emixing proportions and the mixing parameters.}

\item{hist}{The style of the histogram. We can choose to obtain two styles of histogram (hist=1 or 2). hist=0 means no histogram.}

\item{comp}{A parameter for the compontent fitted density. 
comp=T means component fitted densities are drawn, and comp=F means no component fitted densities.}

\item{k}{A single number giving the number of cells for the histogram.}

\item{h}{A number between 0 and 1, specifying the height of the density function, default 
value is 1 (complete density function).}

\item{main}{The title of the graph.}

\item{xlab}{The label for the x-axis}

\item{ylab}{The label for the y-axis}
}
\value{

}
\description{
plotmix.norm0
}
