% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmix.pois.R
\name{plotmix.pois}
\alias{plotmix.pois}
\title{plotmix.pois}
\usage{
plotmix.pois(
  x = NULL,
  x.max = NULL,
  alpha,
  mu,
  sub.pmf = T,
  nclass = NULL,
  main = "",
  xlab = "Counts",
  ylab = "Prob"
)
}
\arguments{
\item{x}{A vector of data whose histogram is to be drawn from.}

\item{x.max}{If x is null, 0:x.max is the range of pmf to be plotted.}

\item{alpha}{A vector of the mixing proportions.}

\item{sub.pmf}{logical, draw subpopulation pmf's.}

\item{nclass}{The number of bars for the histogram.}

\item{main}{The title for the histogram.}

\item{xlab}{The name of the x-axis label.}

\item{ylab}{The name of the y-axis label.}

\item{theta}{A vector of the subpopulation means.}
}
\description{
This function plots the data in histogram with 
             the corresponding fitted mixture pmf
             or the mixture pmf only, for Poisson mixture. 
             The pmf is given as if it is a continuous curve.
}
