% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmix.beta.R
\name{pmix.beta}
\alias{pmix.beta}
\title{pmix.beta}
\usage{
pmix.beta(x, mix_prop, alpha, beta)
}
\arguments{
\item{x}{The cdf value of the beta mixture at x.}

\item{mix_prop}{A vector of the mixing proportions.}

\item{alpha}{A vector of the component alphas.}

\item{beta}{A vector of the component betas.}
}
\description{
pmix.beta
}
\examples{
x = c(rbeta(50,1,10),rbeta(50,9,0.5))
pmix.beta(x,c(.5,.5),c(1,9),c(10,0.5))
}
