% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmix.norm.R
\name{pmix.norm}
\alias{pmix.norm}
\title{pmix.norm}
\usage{
pmix.norm(x, alpha, mu, sigma)
}
\arguments{
\item{x}{The cdf value of the Normal mixture at x.}

\item{alpha}{A vector of the mixing proportions.}

\item{mu}{A vector of the component means.}

\item{sigma}{A vector of the component standard deviations.}
}
\value{

}
\description{
pmix.norm
}
