% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.beta.R
\name{pmle.beta}
\alias{pmle.beta}
\title{pmle.beta}
\usage{
pmle.beta(
  x,
  m0,
  n.iter = 10,
  max.iter = 5000,
  tol = 1e-06,
  epsilon = 1,
  an = NULL,
  seed = NULL,
  maxit = 5000
)
}
\arguments{
\item{x}{A vector of the observed values.}

\item{m0}{The order of the finite mixture model.}

\item{n.iter}{The number of EM iterations for each initial value. The one that gained the most in likelihood will be iterative further.}

\item{max.iter}{The maximum number of iterations for the EM algorithm.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm, default value: tol = 1e-6.}

\item{epsilon}{The size of the penalized function of the mixing distribution, default value: epsilon = 1.}

\item{an}{A size control parameter that controls the severity of the penalty. The recommended value is n^{-3/2}.}

\item{seed}{Used for reproducing the same sequence of output numbers. Default is NULL.}

\item{maxit}{The maximum number of iterations for the Broyden or Newton methods used in the method of moments estimation.}
}
\value{
The PMLE of the parameters with order = m0 (mixing proportions, mixing alphas
         and mixing betas), log-likelihood value at the PMLE and the penalized log-likelihood
         value at the PMLE, classification on which subpopulation does the observed value belong to.
}
\description{
This function computes the PMLE of the parameters under a mixture of betas.
}
\examples{
 
data <- c(rbeta(50,10,2),rbeta(50,3,18))
pmle.beta(data,2)
}
