% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.beta.R
\name{pmle.beta}
\alias{pmle.beta}
\title{pmle.beta}
\usage{
pmle.beta(
  x,
  m0,
  n.init = 10,
  n.iter = 10,
  max.iter = 5000,
  tol = 1e-06,
  epsilon = 1,
  an = NULL,
  maxit = 5000
)
}
\arguments{
\item{x}{A numeric vector of observed values.}

\item{m0}{The number of components (order) in the beta mixture model.}

\item{n.init}{A computer generated n.init initials value.}

\item{n.iter}{The number of EM iterations to perform for each initial value. The initialization 
yielding the highest penalized log-likelihood will be further optimized.}

\item{max.iter}{The maximum number of iterations allowed in the final EM optimization phase.}

\item{tol}{The tolerance threshold for convergence based on change in penalized log-likelihood. 
Default is \code{1e-6}.}

\item{epsilon}{A regularization parameter controlling the smoothing in the E-step. Default is \code{1}.}

\item{an}{A size control parameter that determines the severity of the penalty. 
The recommended value is \eqn{n^{-3/2}}. If \code{NULL}, it is set to \code{length(x)^(3/2)}.}

\item{maxit}{The maximum number of iterations for the nonlinear solvers used in the 
method of moments estimation during initialization.}
}
\value{
A list with the following components:
\item{mix_prop}{Estimated mixing proportions.}
\item{alpha}{Estimated \eqn{\alpha} parameters for each component.}
\item{beta}{Estimated \eqn{\beta} parameters for each component.}
\item{loglik}{Log-likelihood at the PMLE.}
\item{ploglik}{Penalized log-likelihood at the PMLE.}
\item{iter.n}{Number of EM iterations performed after initialization.}
\item{classification}{Component assignment for each observation based on posterior probabilities.}
}
\description{
This function computes the penalized maximum likelihood estimate (PMLE) of the 
parameters under a finite mixture of beta distributions.
}
\examples{
data <- c(rbeta(50, 10, 2), rbeta(50, 3, 18))
pmle.beta(data, 2)
}
