% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.beta.sub.R
\name{pmle.beta.sub}
\alias{pmle.beta.sub}
\title{pmle.beta.sub}
\usage{
pmle.beta.sub(x, m0, para0, an, epsilon)
}
\arguments{
\item{x}{A numeric vector of observed values assumed to follow a beta mixture distribution.}

\item{m0}{The number of components in the finite mixture model.}

\item{para0}{A numeric vector containing initial values for mixing proportions and 
the \eqn{\alpha} and \eqn{\beta} parameters for all components.}

\item{an}{A penalty control parameter, usually chosen as \eqn{n^{-3/2}}, where \eqn{n} is the sample size.}

\item{epsilon}{A smoothing parameter for the mixing proportions in the E-step. Default is \code{1}.}
}
\value{
A numeric vector containing updated estimates of the mixing proportions, 
        \eqn{\alpha} and \eqn{\beta} values for each component, the log-likelihood, 
        and the penalized log-likelihood.
}
\description{
A sub-function for \code{pmle.beta}, performs the core computation for 
             penalized maximum likelihood estimation (PMLE) of the beta mixture model.
}
