% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.beta.sub.R
\name{pmle.beta.sub}
\alias{pmle.beta.sub}
\title{pmle.beta.sub}
\usage{
pmle.beta.sub(x, m0, para0, an, epsilon)
}
\description{
A sub function for pmle.beta, does the actual work of PMLE of the beta mixture.
It is used in the pmle.beta function.
}
