% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.binom.sub.R
\name{pmle.binom.sub}
\alias{pmle.binom.sub}
\title{pmle.binom.sub}
\usage{
pmle.binom.sub(
  count,
  freq,
  size,
  m0,
  lambda,
  init.val,
  n.init,
  max.iter,
  n.iter,
  tol
)
}
\description{
A sub function for pmle.binom, does the actual work of PMLE binomial mixture.
It is used in the pmle.binom function.
}
