% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.exp.sub.R
\name{pmle.exp.sub}
\alias{pmle.exp.sub}
\title{pmle.exp.sub}
\usage{
pmle.exp.sub(xx, m0, lambda, init.val, n.init, n.iter, max.iter, tol)
}
\description{
This function does the real work of computing the PMLE of the mixing distribution under the exponential mixture model. 
It is used in the pmle.exp function.
}
