% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.norm.R
\name{pmle.norm}
\alias{pmle.norm}
\title{pmle.norm}
\usage{
pmle.norm(
  x,
  m0 = 1,
  lambda = 0,
  inival = NULL,
  len = 10,
  niter = 50,
  tol = 1e-06,
  rformat = FALSE
)
}
\arguments{
\item{x}{The data that can either be a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequencies.}

\item{m0}{The order of the finite mixture model, default value: m0 = 1.}

\item{lambda}{The size of the penalized function of the mixing distribution, default value: lambda = 0.}

\item{inival}{The initial values chosen for the EM-algorithm, a 3m0-dimension vector including m0 mixing proportions, 
m0 component means and m0 component variances, or a matrix with 3m0 columns, default value: inival = NULL. (if not provided, random initial values are used.)}

\item{len}{The number of initial values chosen for the EM-algorithm, default value: len = 10.}

\item{niter}{The smallest number of iterations required for all initial values in the EM-algorithm. 
The algorithm runs EM-iteration niter times from each initial value. 
The iteration will restart from the parameter value with the highest likelihood value at the point and run until convergence. default value: niter = 50.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm, default value: tol = 1e-6.}

\item{rformat}{The format of the output. If rformat=T, 
it means the output format is determined by R software. 
If rformat=F, it means the output format is determined by our default setting. 
When the output is larger than 0.001, it is determined by round(output,3); 
When the output is less than 0.001, it is determined by signif(output,3).
The default value of rformat is F.}
}
\value{
The PMLE or MLE of the parameters with order = m0 (mixing proportions, mixing means
}
\description{
This function computes the PMLE or MLE of the parameters under a mixture of normals with unequal variance.
}
\examples{
#load the pearson's crab data, fit the 2 and 3 component normal mixture models,
plot the histgorams of the observations and the fitted densities.
data(pearson)
out1 <- pmle.norm(pearson,2,1)
plotmix.norm(pearson,out1)

# Not run:
out2 <- pmle.norm(pearson,3,1)
plotmix.norm(pearson,out2)
par(mfrow=c(1,1))

# End(Not run)
}
