% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.norm.sub.a.R
\name{pmle.norm.sub.a}
\alias{pmle.norm.sub.a}
\title{pmle.norm.sub.a}
\usage{
pmle.norm.sub.a(xx, sample.var, m0, para0, lambda, an)
}
\arguments{
\item{xx}{The input data.}

\item{sample.var}{The sample variance of the input data.}

\item{m0}{The order of the finite mixture model.}

\item{para0}{The estimated mixing parameters.}

\item{lambda}{The size of the penalized function of the mixing distribution, default value: lambda = 1.}

\item{an}{A penalty on the variance(ChenTanZhangSinica2008). The recommended value is n^{-1/2}.}
}
\description{
It is used in the pmle.norm.sub function, it computes a single EM iteration for the univariate Gaussian mixture.
}
