% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.norm.sub.a.R
\name{pmle.norm.sub.a}
\alias{pmle.norm.sub.a}
\title{pmle.norm.sub.a}
\usage{
pmle.norm.sub.a(xx, sample.var, m0, para0, lambda, an)
}
\description{
It is used in the pmle.norm.sub function, it computes a single EM iteration for the univariate Gaussian mixture.
}
