% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.norm.sub.a.R
\name{pmle.norm.sub.a}
\alias{pmle.norm.sub.a}
\title{pmle.norm.sub.a}
\usage{
pmle.norm.sub.a(x, m, para0, lambda, an)
}
\arguments{
\item{x}{The data that can either be a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequencies.}

\item{m}{The order of the mixture}

\item{para0}{}

\item{lambda}{The size of the penalized function of the mixing distribution, default value: lambda = 0.}

\item{an}{}
}
\description{
pmle.norm.sub.a
}
