% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.norm.sub.R
\name{pmle.norm.sub}
\alias{pmle.norm.sub}
\title{pmle.norm.sub}
\usage{
pmle.norm.sub(x, m0, lambda, an, init.val, n.init, n.iter, max.iter, tol)
}
\description{
It is used in the pmle.norm function, it does the actual EM-algorithm and computes the penalized MLE.
}
