% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.norm.sub.R
\name{pmle.norm.sub}
\alias{pmle.norm.sub}
\title{pmle.norm.sub}
\usage{
pmle.norm.sub(x, m0, lambda, an, init.val, n.init, n.iter, max.iter, tol)
}
\arguments{
\item{x}{The data that can either be a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequencies.}

\item{m0}{The order of the finite mixture model, default value: m0 = 1}

\item{lambda}{The size of the penalized function of the mixing distribution, default value: lambda = 0.}

\item{an}{}

\item{init.val}{The initial values chosen for the EM-algorithm, a 3m0-dimension vector including m0 mixing proportions, 
m0 component means and m0 component variances, or a matrix with 3m0 columns, 
default value: inival = NULL. (if not provided, random initial values are used.)}

\item{n.init}{}

\item{n.iter}{The smallest number of iterations required for all initial values in the EM-algorithm.}

\item{max.iter}{The maximum amount of iterations.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm, default value: tol = 1e-6.}
}
\value{

}
\description{
pmle.norm.sub
}
