% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.norm0.R
\name{pmle.norm0}
\alias{pmle.norm0}
\title{pmle.norm0}
\usage{
pmle.norm0(
  x,
  var.sub = 1,
  m0,
  lambda = 1,
  init.val = NULL,
  n.init = 10,
  n.iter = 50,
  max.iter = 5000,
  tol = 1e-08,
  rformat = FALSE
)
}
\arguments{
\item{x}{The input data that can either be a vector or a matrix with the 1st column being the observed values
and the 2nd column being the corresponding frequencies.}

\item{var.sub}{A known subpopulation variance, default at 1.}

\item{m0}{The order of the finite mixture model.}

\item{lambda}{The size of the penalty function of the mixing proportions.}

\item{n.init}{The    number of initial values for the EM-algorithm.}

\item{n.iter}{The number of EM iterations for all initial values.}

\item{max.iter}{Maximum amount of EM iterations, it stops at 5000.}

\item{tol}{The tolerance value for the convergence of the EM-algorithm, default value: tol = 1e-8.}

\item{rformat}{The format of the output. If rformat=T, 
it means the output format is determined by R software. 
If rformat=F, it means the output format is determined by our default setting. 
When the output is larger than 0.001, it is determined by round(output,3); 
When the output is less than 0.001, it is determined by signif(output,3).
The default value of rformat is F.}

\item{ini.val}{The initial values chosen for the EM-algorithm, a 3m0-dimension vector including m0 mixing proportions, 
m0 component means and m0 component variances, or a matrix with 3m0 columns, default value: inival = NULL. (if not provided, random initial values are used.)}

\item{iter.n}{This reports the number of EM-iteration employed.}
}
\value{
It returns the PMLE or MLE of the parameters with order = m0 (mixing proportions and component parameters), 
log-likelihood value at the PMLE or MLE and the penalized log-likelihood value at the PMLE.
}
\description{
This function computes the PMLE or MLE of the parameters under a mixture of normals with equal and known
variance. When the level of penalty is 0, PMLE reduces to MLE.
}
\examples{
x <- rmix.norm(200,c(0.3,0.7),c(-1,2))
pmle.norm0(x,m0=2)
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
