% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.norm0.sub.R
\name{pmle.norm0.sub}
\alias{pmle.norm0.sub}
\title{pmle.norm0.sub}
\usage{
pmle.norm0.sub(x, m0, lambda, init.val, n.init, n.iter, max.iter, tol)
}
\description{
It is used in the pmle.norm0 function, It does the actual computing for the EM-algorithm.
}
