% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.pois.R
\name{pmle.pois}
\alias{pmle.pois}
\title{pmle.pois}
\usage{
pmle.pois(
  x,
  m0,
  lambda = 1,
  init.val = NULL,
  n.init = 10,
  n.iter = 50,
  tol = 1e-06,
  max.iter = 5000,
  rformat = FALSE
)
}
\arguments{
\item{x}{Data that can be either a vector or a matrix of observations. 
1st column: counts 
2nd column: frequencies.}

\item{m0}{The order of the finite mixture model.}

\item{lambda}{The size of the penalty function of the mixing proportions.}

\item{init.val}{A user provided initial values for the EM-algorithm to 
compute the PMLE under the null model.}

\item{n.init}{The number of initial values chosen for the EM-algorithm.}

\item{n.iter}{Least amount of iterations for all initial values in the EM-algorithm.}

\item{max.iter}{Maximum amount of iterations.}

\item{rformat}{A specific format, please see rousignif.R function.}
}
\description{
This function computes the PMLE of parameters under a Poisson mixture
}
