% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle.pois.sub.R
\name{pmle.pois.sub}
\alias{pmle.pois.sub}
\title{pmle.pois.sub}
\usage{
pmle.pois.sub(x, m0, lambda, init.val, n.init, n.iter, tol, max.iter = 5000)
}
\description{
This function is used by pmle.pois. It does the real work but not stand alone.
}
