% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmix.exp.R
\name{qmix.exp}
\alias{qmix.exp}
\title{qmix.exp}
\usage{
qmix.exp(qq, alpha, mu)
}
\arguments{
\item{qq}{The level of the quantile to be computed.}

\item{alpha}{A vector of the mixing probabilities.}

\item{mu}{A vector of the subpopulation means.}
}
\description{
This function computes the quantile of the finite exponential mixture distribution
}
