% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmix.exp.R
\name{qmix.exp}
\alias{qmix.exp}
\title{qmix.exp}
\usage{
qmix.exp(alpha, theta, qq)
}
\arguments{
\item{alpha}{A vector of the mixing probabilities.}

\item{theta}{A vector of the subpopulation means.}

\item{qq}{The level of the quantile to be computed.}
}
\description{
This function computes the quantile of the finite exponential mixture distribution
}
