% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmix.norm.R
\name{qmix.norm}
\alias{qmix.norm}
\title{qmix.norm}
\usage{
qmix.norm(pp, alpha, mu, sigma, tol = 1e-08)
}
\description{
This function computes the quantile of the univariate finite normal mixture one by one.
It is used in the tildeB22.norm0 function.
}
