% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmix.norm.R
\name{qmix.norm}
\alias{qmix.norm}
\title{qmix.norm}
\usage{
qmix.norm(pp, alpha, mu, sigma, tol = 1e-08)
}
\arguments{
\item{pp}{}

\item{alpha}{A vector of the mixing proportions.}

\item{mu}{A vector of the component means.}

\item{sigma}{A vector of the component standard deviations.}

\item{tol}{}
}
\description{
qmix.norm
}
