% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmix.exp.R
\name{rmix.exp}
\alias{rmix.exp}
\title{rmix.exp}
\usage{
rmix.exp(n, alpha, theta)
}
\arguments{
\item{n}{The sample size.}

\item{alpha}{A vector of the mixing probabilities.}

\item{theta}{A vector of the subpopulation means.}
}
\value{
It returns a sample of size n from an m-component exponential mixture.
}
\description{
This function generates the iid samples from the finite exponential mixture
}
\examples{
n = 4000
mu = c(3, 9, 18)
alpha = c(.5, .3, .2)
rmix.exp(n, alpha, mu)
}
\author{
Shaoting Li, Jiahua Chen and Pengfei Li
}
