% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmix.exp.R
\name{rmix.exp}
\alias{rmix.exp}
\title{rmix.exp}
\usage{
rmix.exp(n, alpha, theta)
}
\arguments{
\item{n}{The sample size.}

\item{alpha}{A vector of the mixing probabilities.}

\item{theta}{A vector of the subpopulation means.}
}
\description{
This function generates the iid samples from the finite exponential mixture
}
