% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmix.pois.R
\name{rmix.pois}
\alias{rmix.pois}
\title{rmix.pois}
\usage{
rmix.pois(n, alpha, theta)
}
\arguments{
\item{n}{The sample size.}

\item{alpha}{A vector of the mixing proportions.}

\item{theta}{A vector of the subpopulation means.}
}
\description{
This function generates a random sample from a finite Poisson mixture
}
\examples{
#Generate a random sample from a 2 component Poisson mixture,
and computes the sample mean and variance. 
x <- rmix.pois(200,c(0.3,0.7),c(2,5))
mean(x)
var(x)
}
