% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rousignif.R
\name{rousignif}
\alias{rousignif}
\title{rousignif}
\usage{
rousignif(x)
}
\description{
This function returns the output combining round and signif functions.
It is used in various main functions in the package.
}
