% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rousignif.R
\name{rousignif}
\alias{rousignif}
\title{rousignif}
\usage{
rousignif(x)
}
\arguments{
\item{x}{A single numeric or a numeric vector or matrix.}
}
\value{

}
\description{
This function return the output combining round and signif functions.
}
