% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar.beta.mix.R
\name{sar.beta.mix}
\alias{sar.beta.mix}
\title{sar.beta.mix}
\usage{
sar.beta.mix(x, m0)
}
\arguments{
\item{x}{A numeric vector of observations in (0, 1).}

\item{m0}{Integer; the number of components in the mixture.}
}
\value{
A list containing:
\item{mix_prop}{Estimated mixing proportions for each component.}
\item{alpha}{Estimated alpha (shape1) parameters.}
\item{beta}{Estimated beta (shape2) parameters.}
}
\description{
A general closed-form estimator for finite beta mixture models, based on the 
             Score-Adjusted and Refined (SAR) method extended to mixtures. This function is also 
             used as an initializer for the EM algorithm in the pmle.beta function.
}
