% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tildeB22.binom.R
\name{tildeB22.binom}
\alias{tildeB22.binom}
\title{tildeB22.binom}
\usage{
tildeB22.binom(alpha, theta, size)
}
\arguments{
\item{alpha}{A vector of the mixing probabilities.}

\item{theta}{A vector of the probabilities of success for each subpopulation.}

\item{size}{The number of trials in the binomial subpopulation.}
}
\description{
This function computes the \tilde B_{22} matrix in the
EM-test for one-parameter subpopulation distributions,
including binomial. See Chen and Li (2010JASA)
}
