% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tildeB22.binom.R
\name{tildeB22.binom}
\alias{tildeB22.binom}
\title{tildeB22.binom}
\usage{
tildeB22.binom(alpha, theta, size)
}
\description{
This function computes the tilde{B}_{22} matrix in the
EM-test for one-parameter subpopulation distributions,
including binomial. See Chen and Li (2010JASA). 
It is used in the emtest.binom function.
}
