% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tildeB22.exp.R
\name{tildeB22.exp}
\alias{tildeB22.exp}
\title{tildeB22.exp}
\usage{
tildeB22.exp(alpha, mu)
}
\description{
This function computes \tilde B_{22} under the Exponential mixture.
It is used in the emtest.exp function.
}
