% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tildeB22.norm0.R
\name{tildeB22.norm0}
\alias{tildeB22.norm0}
\title{tildeB22.norm0}
\usage{
tildeB22.norm0(alpha, theta, N = 10000)
}
\description{
This function computes tilde{B}_{22} matrix in the
             EM-test for one-parameter subpopulation distributions including normal with known variance.
             It is used in the emtest.norm0 function.
}
