% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tildeB22.norm0.R
\name{tildeB22.norm0}
\alias{tildeB22.norm0}
\title{tildeB22.norm0}
\usage{
tildeB22.norm0(alpha, theta, N = 10000)
}
\arguments{
\item{alpha}{A vector of the mixing proportions.}

\item{theta}{A vector of the subpopulation means.}

\item{N}{}
}
\description{
This function computes \tilde B_{22} matrix in the
             EM-test for one-parameter subpopulation distributions
             including normal with known variance.
}
\note{
See Chen and Li (2010JASA).
      \tilde B_{22} in this R-function is the correlation matrix.
      Renamed from tb2.
}
