% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tildeB22.pois.R
\name{tildeB22.pois}
\alias{tildeB22.pois}
\title{tildeB22.pois}
\usage{
tildeB22.pois(alpha, theta)
}
\arguments{
\item{alpha}{A vector of the mixing proportions.}

\item{theta}{A vector of the subpopulation means.}
}
\description{
This function computes \tilde B_{22} under the Poisson mixture.
             \tilde B_{22} is standardized to the correlation matrix.
}
