% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tildeB22.pois.R
\name{tildeB22.pois}
\alias{tildeB22.pois}
\title{tildeB22.pois}
\usage{
tildeB22.pois(alpha, theta)
}
\description{
This function computes tilde{B}_{22} under the Poisson mixture.
             tilde{B}_{22} is standardized to the correlation matrix.
             It is used in the emtest.pois function.
}
